Support older ES client versions
