+++ apm-agent-plugins/apm-es-restclient-plugin/apm-es-restclient-plugin-5_6/pom.xml
@@ -1,0 +1,43 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <artifactId>apm-es-restclient-plugin</artifactId>
+        <groupId>co.elastic.apm</groupId>
+        <version>0.9.0-SNAPSHOT</version>
+    </parent>
+
+    <artifactId>apm-es-restclient-plugin-5_6</artifactId>
+    <name>${project.groupId}:${project.artifactId}</name>
+
+    <properties>
+        <version.elasticsearch>5.6.0</version.elasticsearch>
+    </properties>
+
+    <dependencies>
+        <!-- Elasticsearch rest client -->
+        <dependency>
+            <groupId>org.elasticsearch.client</groupId>
+            <artifactId>elasticsearch-rest-client</artifactId>
+            <version>${version.elasticsearch}</version>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.elasticsearch.client</groupId>
+            <artifactId>elasticsearch-rest-high-level-client</artifactId>
+            <version>${version.elasticsearch}</version>
+            <scope>provided</scope>
+        </dependency>
+
+        <!-- Elasticsearch testcontainer -->
+        <dependency>
+            <groupId>fr.pilato.elasticsearch.testcontainers</groupId>
+            <artifactId>testcontainers-elasticsearch</artifactId>
+            <version>0.1</version>
+            <scope>test</scope>
+        </dependency>
+    </dependencies>
+</project>
