+++ apm-agent-plugins/apm-es-restclient-plugin/apm-es-restclient-plugin-6_4/src/main/resources/META-INF/services/co.elastic.apm.bci.ElasticApmInstrumentation
@@ -1,0 +1,1 @@
+co.elastic.apm.es.restclient.v6_4.ElasticsearchRestClientInstrumentation
