+++ apm-agent-plugins/apm-es-restclient-plugin/pom.xml
@@ -8,36 +8,14 @@
         <version>0.9.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
+    <packaging>pom</packaging>
+
+    <modules>
+        <module>apm-es-restclient-plugin-5_6</module>
+        <module>apm-es-restclient-plugin-6_4</module>
+    </modules>
 
     <artifactId>apm-es-restclient-plugin</artifactId>
     <name>${project.groupId}:${project.artifactId}</name>
-
-    <properties>
-        <version.elasticsearch>6.4.1</version.elasticsearch>
-    </properties>
-
-    <dependencies>
-        <!-- Elasticsearch rest client -->
-        <dependency>
-            <groupId>org.elasticsearch.client</groupId>
-            <artifactId>elasticsearch-rest-client</artifactId>
-            <version>${version.elasticsearch}</version>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.elasticsearch.client</groupId>
-            <artifactId>elasticsearch-rest-high-level-client</artifactId>
-            <version>${version.elasticsearch}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- Elasticsearch testcontainer -->
-        <dependency>
-            <groupId>fr.pilato.elasticsearch.testcontainers</groupId>
-            <artifactId>testcontainers-elasticsearch</artifactId>
-            <version>0.1</version>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
 
 </project>
