--- apm-agent-plugins/apm-es-restclient-plugin/src/main/resources/META-INF/services/co.elastic.apm.bci.ElasticApmInstrumentation
@@ -1,1 +1,0 @@
-co.elastic.apm.es.restclient.ElasticsearchRestClientInstrumentation
