+++ elastic-apm-agent/pom.xml
@@ -173,7 +173,12 @@
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>apm-es-restclient-plugin</artifactId>
+            <artifactId>apm-es-restclient-plugin-5_6</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>apm-es-restclient-plugin-6_4</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
